split the circular dependency problem
