+++ src/tcg/decks/UbelDeck.ts
@@ -6,6 +6,7 @@ import { CardEmoji } from "../formatting/emojis";
 import { MessageCache } from "../../tcgChatInteractions/messageCache";
 import { TCGThread } from "../../tcgChatInteractions/sendGameMessage";
 import { signatureMoves } from "./utilDecks/signatureMoves"
+import { a_malevolentShrine} from "./utilDecks/ubelSignature"
 
 const a_reelseiden = new Card({
   title: "Reelseiden",
@@ -97,38 +98,7 @@ const a_dismantle = new Card({
   },
 });
 
-export const a_malevolentShrine = new Card({
-  title: "Malevolent Shrine",
-  description: ([dmg]) =>
-    `HP-11. Has a 80% of missing if the opponent didn't attack last turn. DMG ${dmg}.`,
-  cosmetic: {
-    cardGif: "https://media.discordapp.net/attachments/1338831179981262943/1363264315272073406/malevolent-shrine-ubel.gif?ex=68060f14&is=6804bd94&hm=300b3e5578f56a069ea858f0f660ce855be6a3f6f32f246b434066ea770da58e&=&width=400&height=225",
-  },
-  emoji: CardEmoji.UBEL_CARD,
-  cardMetadata: { nature: Nature.Attack, signature: true, ubelFailureRate: 80 },
-  effects: [22],
-  hpCost: 11,
-  cardAction: function (
-    this: Card,
-    game,
-    characterIndex,
-    messageCache: MessageCache
-  ) {
-    const character = game.getCharacter(characterIndex);
-    const opponent = game.getCharacter(1 - characterIndex);
-    const pierceFactor = (character.additionalMetadata.pierceFactor ??= 0);
-    messageCache.push(
-      `# ጠ ል ረ ቿ ሀ ዐ ረ ቿ ክ ፕ    ነ ዘ ዪ ጎ ክ ቿ!`,
-      TCGThread.Gameroom
-    );
 
-    CommonCardAction.pierceAttack(game, characterIndex, {
-      damage: this.calculateEffectValue(this.effects[0]),
-      hpCost: this.hpCost,
-      pierceFactor: pierceFactor,
-    });
-  },
-});
 
 export const rushdown = new Card({
   title: "Rushdown",
@@ -157,6 +127,7 @@ export const rushdown = new Card({
         description: `Increases SPD by ${spdIncrease} for ${turnCount} turns. Attacks will not miss`,
         turnDuration: turnCount,
         tags: { ubelSpeedModifiers: 1 },
+        
         endOfTurnAction: (_game, _characterIndex, _messageCache) => {
           messageCache.push(
             `${character.name} is being reckless.`,
@@ -164,6 +135,7 @@ export const rushdown = new Card({
           );
           character.adjustStat(-5, StatsEnum.HP);
         },
+
         endOfTimedEffectAction: (_game, _characterIndex, _messageCache) => {
           _messageCache.push(`${character.name} retreats.`, TCGThread.Gameroom);
           character.adjustStat(-1 * spdIncrease, StatsEnum.SPD);
@@ -186,14 +158,14 @@ const recompose = new Card({
   effects: [10],
   cardAction: function (this: Card, game, characterIndex, messageCache) {
     const character = game.characters[characterIndex];
-    messageCache.push(`${character.name} calms down.`, TCGThread.Gameroom);
-
-    character.adjustStat(-10, StatsEnum.SPD);
-    character.adjustStat(
-      this.calculateEffectValue(this.effects[0]),
-      StatsEnum.HP
+    messageCache.push(`${character.name} takes cover to ponder the fleeting nature of her life.`,
+      TCGThread.Gameroom
     );
+
     const turnCount = 2;
+    character.adjustStat(-10, StatsEnum.SPD);
+    const hpIncrease = this.calculateEffectValue(this.effects[0])
+    character.adjustStat(hpIncrease,StatsEnum.HP);
 
     CommonCardAction.replaceOrAddNewTimedEffect(
       game,
@@ -204,6 +176,7 @@ const recompose = new Card({
         description: `Decreases SPD by 10 for ${turnCount} turns. Attacks will not hit`,
         turnDuration: turnCount,
         tags: { ubelSpeedModifiers: 1 },
+
         endOfTurnAction: (_game, _characterIndex, _messageCache) => {
           messageCache.push(
             `${character.name} took a break and recoups.`,
@@ -211,6 +184,7 @@ const recompose = new Card({
           );
           character.adjustStat(this.effects[0]/2, StatsEnum.HP);
         },
+
         endOfTimedEffectAction: (_game, _characterIndex, _messageCache) => {
           messageCache.push(
             `${character.name} has recomposed ${character.cosmetic.pronouns.reflexive}.`,
@@ -223,6 +197,8 @@ const recompose = new Card({
         },
       })
     );
+
+    character.timedEffects.push();
   },
 });
 
@@ -290,6 +266,7 @@ export const sorganeil = new Card({
   },
 });
 
+/*
 export const empathy = new Card({
   title: `Empathy`,
   cardMetadata: { nature: Nature.Util },
@@ -340,8 +317,9 @@ export const empathy = new Card({
     }
   },
 });
+*/
 
-export const imitate = new Card({
+export const empathy = new Card({
   title: "Empathy",
   cardMetadata: { nature: Nature.Util },
   description: () =>
